==> Credit Card Transaction Fraud Analytics/1. Clean Data.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_excel('cleaning data_3.xlsx', converters={'Merchnum': lambda x: str(x), 'Merch_zip': lambda x : str(x)})

df.info()

## 1. Treat Merchnum == 0 as nan

df['Merchnum'] = df['Merchnum'].replace('0',np.nan)

df[df['Merchnum'] == '0']

## 2. Filling in missing value

group_merchnum_1 = df.groupby('Merchnum')['Merch_zip'].apply(pd.Series.mode).to_frame().reset_index().groupby('Merchnum').agg({'Merch_zip':'first'})
group_merchnum_2 = df.groupby('Merchnum')['Merch_state'].apply(pd.Series.mode).to_frame().reset_index().groupby('Merchnum').agg({'Merch_state':'first'})

group_merchnum = df['Merchnum'].unique()
group_merchnum = pd.DataFrame({'Merchnum':group_merchnum})
group_merchnum = group_merchnum.set_index('Merchnum')

group_merchnum = pd.merge(group_merchnum, group_merchnum_1,left_index = True, right_index = True,how = 'left')
group_merchnum = pd.merge(group_merchnum, group_merchnum_2,left_index = True, right_index = True,how = 'left')

group_merchnum

group_description_1 = df[~df['Merch_description'].str.contains('ADJ')].groupby('Merch_description')['Merch_zip'].apply(pd.Series.mode).to_frame().reset_index().groupby('Merch_description').agg({'Merch_zip':'first'})
group_description_2 = df[~df['Merch_description'].str.contains('ADJ')].groupby('Merch_description')['Merch_state'].apply(pd.Series.mode).to_frame().reset_index().groupby('Merch_description').agg({'Merch_state':'first'})
group_description_3 = df[~df['Merch_description'].str.contains('ADJ')].groupby('Merch_description')['Merchnum'].apply(pd.Series.mode).to_frame().reset_index().groupby('Merch_description').agg({'Merchnum':'first'})

group_description_4 = df[~df['Merch_description'].str.contains('ADJ')].groupby('Merch_description').agg({'Recnum':'first'})

group_description = df[~df['Merch_description'].str.contains('ADJ')]
group_description = group_description['Merch_description'].unique()
group_description = pd.DataFrame({'Merch_description':group_description})
group_description = group_description.set_index('Merch_description')

group_description = pd.merge(group_description,group_description_1,left_index = True, right_index = True, how = 'left')
group_description = pd.merge(group_description,group_description_2,left_index = True, right_index = True, how = 'left')
group_description = pd.merge(group_description,group_description_3,left_index = True, right_index = True, how = 'left')
group_description = pd.merge(group_description,group_description_4,left_index = True, right_index = True, how = 'left')

group_description['Merch_zip'] = group_description['Merch_zip'].fillna(group_description['Recnum'])
group_description['Merch_state'] = group_description['Merch_state'].fillna(group_description['Recnum'])

group_description

group_description.loc['AOAC INTERNATIONAL']

df.head()

### 2.1 Fill state and zip by merchnum group

df = df.set_index('Merchnum')
df['Merch_state'] = df['Merch_state'].fillna(group_merchnum['Merch_state'])
df['Merch_state'] = df['Merch_state'].replace('nan', np.nan)
df['Merch_state'].count()

df['Merch_zip'] = df['Merch_zip'].fillna(group_merchnum['Merch_zip'])
df['Merch_zip'] = df['Merch_zip'].replace('nan', np.nan)
df = df.reset_index()
df['Merch_zip'].count()

### 2.1 Fill state, zip and merchnum by merch_description group

df = df.set_index('Merch_description')
df['Merch_state'] = df['Merch_state'].fillna(group_description['Merch_state'])
df['Merch_state'] = df['Merch_state'].replace('nan', np.nan)
df['Merch_state'].count()

df['Merch_zip'] = df['Merch_zip'].fillna(group_description['Merch_zip'])
df['Merch_zip'] = df['Merch_zip'].replace('nan', np.nan)
df['Merch_zip'].count()

df['Merchnum'] = df['Merchnum'].fillna(group_description['Merchnum'])
df['Merchnum'] = df['Merchnum'].replace('nan', np.nan)
df['Merchnum'].count()

df = df.reset_index()

df.info()

### 2.3 Filling in missing zip and state using minus recordnum

df['minus'] = -df['Recnum']

df['minus'] = df['minus'].astype('str')

df['Merch_state'] = df['Merch_state'].fillna(df['minus'])
df['Merch_zip'] = df['Merch_zip'].fillna(df['minus'])

df.info()

### 2.4 assign merchnum by grouping by merch_description & state & zip

df['Merch_state'] = df['Merch_state'].astype('str')
df['Merch_zip'] = df['Merch_zip'].astype('str')

df['assign'] = df['Merch_description'] + df['Merch_state'] + df['Merch_zip']

merchnum_missing = df.groupby('assign').agg({'minus':'first'})

df = df.set_index('assign')
df['Merchnum'] = df['Merchnum'].fillna(merchnum_missing['minus'])
df = df.reset_index()
df.info()

df.head()

df.to_excel('no_missing.xlsx')

df[df['Merch_description'] == 'AOAC INTERNATIONAL']

df[df['Merch_description'].str.contains('GRAINGER')].tail(20)

df[df['Merch_description'].str.contains('CELLULAR ONE')].tail(20)
