==> Credit Card Transaction Fraud Analytics/2. Variable Creation - Amount_variables.ipynb <==
import pandas as pd
import numpy as np


# data preparing

df = pd.read_excel('no_missing.xlsx', converters={'Merchnum': lambda x: str(x), 'Merch_zip': lambda x : str(x)})
df = df.drop(['Unnamed: 0','assign','minus'],axis = 1)
df.info()

for x in df.columns:
    if x != "Date" and x!="Amount" and x!="Recnum":
        df[x] = df[x].astype(str)
df.info()

# create attributes

# Cardnum
# Merchnum
# Merch description(part)
# Cardnum-Merchnum
# Cardnum-Merch description(part)
# Cardnum-Merch state
# Cardnum-Merch zip
# Cardnum-Merch state-Merch zip

df["cardnum-merchnum"] = df["Cardnum"]+df["Merchnum"]
df["cardnum-merch_description_part"] = df["Cardnum"] + df["Merch_description_part"]
df["cardnum-merch_state"] = df["Cardnum"] + df["Merch_state"]
df["cardnum-merch_zip"] = df["Cardnum"] + df["Merch_zip"]
df["cardnum-merch_state_merch_zip"] = df["Cardnum"] + df["Merch_state"] + df["Merch_zip"]

attributes = ["Cardnum","Merchnum","Merch_description_part","cardnum-merchnum","cardnum-merch_description_part","cardnum-merch_state","cardnum-merch_zip","cardnum-merch_state_merch_zip"]

df.head()

# import datetime
# df['join_ts1'] = df['Date']
# for t in [0,1,3,7,14,30]:
#     end_time = 'join_ts2_'+str(t)
#     df[end_time] = df['Date'] + datetime.timedelta(t)

df

import datetime
data1 = df.copy()
data2 = df.copy()
data2['join_ts1'] = data2['Date']
for t in [0,1,3,7,14,30]:
    end_time = 'join_ts2_'+str(t)
    data2[end_time] = data2['Date'] + datetime.timedelta(t)
    
# data3 = data1[['Recnum','Date', 'Cardnum']]
# data4 = data2[['join_ts1','join_ts2_0','join_ts2_1','join_ts2_3','join_ts2_7','join_ts2_30','Cardnum',"Recnum"]]

new_df = pd.DataFrame()

for entity in attributes:
    data3 = data1[['Recnum','Date', entity]]
    data4 = data2[['join_ts1','join_ts2_0','join_ts2_1','join_ts2_3','join_ts2_7','join_ts2_14','join_ts2_30',entity,'Recnum']]
    tempdf = pd.merge(data3,data4, left_on = entity, right_on = entity)
    for t in [0,1,3,7,14,30]:
        name_end_time = 'join_ts2_' + str(t)
        tempdf2 = tempdf[(tempdf['Date'] >= tempdf['join_ts1']) & (tempdf['Date'] <= tempdf[name_end_time])& (tempdf['Recnum_x']>=tempdf['Recnum_y'])]
        tempdf2 = tempdf2.set_index("Recnum_y")
        amount_tempt = data1[["Amount","Recnum"]].set_index("Recnum")
        tempdf3 = pd.merge(tempdf2,amount_tempt,left_index = True, right_index = True).reset_index()
        tempdf3 = tempdf3.rename(columns = {"index":"Amount_Rec","Recnum_x":"Recnum"})
        
        temp_name = str(entity) + '_' + 'count' + str(t) + '_'
        
        tempGroupby = tempdf3[["Recnum","Amount"]].groupby("Recnum").agg(["sum","mean","max","median"])
        tempGroupby.columns = tempGroupby.columns.droplevel(0)
        
        tempGroupby = pd.merge(tempGroupby,amount_tempt,left_index = True, right_index = True, how = "left")
        
        tempGroupby[temp_name+"_actual_sum"] = tempGroupby["Amount"]/tempGroupby["sum"]
        tempGroupby[temp_name+"_actual_avg"]= tempGroupby["Amount"]/tempGroupby["mean"]
        tempGroupby[temp_name+"_actual_max"]= tempGroupby["Amount"]/tempGroupby["max"]
        tempGroupby[temp_name+"_actual_median"]= tempGroupby["Amount"]/tempGroupby["median"]
        
        tempGroupby = tempGroupby.rename(columns = {"sum":temp_name+"_sum","mean":temp_name+"_mean","max":temp_name+"_max","median":temp_name+"_median"})
        tempGroupby.drop(columns=["Amount"])
        new_df = pd.merge(tempGroupby,new_df,left_index = True, right_index = True, how = 'left')
#         tempGroupby = getattr(tempGroupby, 'count')().add_prefix(temp_name)
#         tempGroupby.index = tempGroupby.index.rename('record')
#         tempGroupby.index = tempGroupby.index.astype('str')
#         finaldf = pd.merge(finaldf, tempGroupby,left_index = True, right_index = True, how = 'left')

new_df.columns

new_df.shape

new_df = new_df.drop(columns=["Amount_x"])

new_df = new_df.drop(columns=["Amount_y"])

new_df

new_df.to_csv("amount_variables.csv")
