==> Credit Card Transaction Fraud Analytics/2. Variable Creation - Daysince & Frequency.ipynb <==
import pandas as pd
import numpy as np
import datetime

df = pd.read_excel('no_missing.xlsx', converters={'Merchnum': lambda x: str(x), 'Cardnum': lambda x: str(x), 'Merch_zip': lambda x : str(x)})

df = df.drop(['Unnamed: 0','assign','minus'],axis = 1)

df.head()

df.info()

## Entities

df['Cardnum-Merchnum'] = df['Cardnum'] + df['Merchnum']
df['Cardnum-Merch_description_part'] = df['Cardnum'] + df['Merch_description_part']
df['Cardnum-Merch_state'] = df['Cardnum'] + df['Merch_state']
df['Cardnum-Merch_zip'] = df['Cardnum'] + df['Merch_zip']
df['Cardnum-Merch_state-Merch_zip'] = df['Cardnum'] + df['Merch_state'] + df['Merch_zip']

entities = list(df.columns)
entities.remove('Recnum')
entities.remove('Date')
entities.remove('Merch_state')
entities.remove('Merch_zip')
entities.remove('Merch_description')
entities.remove('Amount')
entities.remove('Fraud')

entities

df.head()

## Days-since Variables

daysince = df[['Recnum']].set_index('Recnum').copy()

for entity in entities:
    data1 = df[['Recnum','Date',entity]]
    data2 = df[['Recnum','Date',entity]]
    data1.Recnum = data1.Recnum.astype(int)
    data2.Recnum = data2.Recnum.astype(int)
    tempdf = pd.merge(data1, data2, left_on = entity, right_on = entity)
    tempdf = tempdf[tempdf['Recnum_x'] > tempdf['Recnum_y']]
    new_col = entity+'_' +'daysince'
    tempdf[new_col] = tempdf['Date_x'] - tempdf['Date_y']
    tempdf[new_col] = tempdf[new_col].astype('timedelta64[D]')
    tempday = tempdf.groupby('Recnum_x').agg({new_col:'min'})
    tempday.index = tempday.index.rename('Recnum')
    daysince = pd.merge(daysince, tempday, left_index = True, right_index = True, how = 'left')

daysince.info()

daysince = daysince.fillna(365)

daysince.info()

daysince.to_csv('Days-since Variables.csv')

## Frequency Variables

frequency = df[['Recnum']].set_index('Recnum').copy()

df['join_ts1'] = df['Date']

for t in [0,1,3,7,14,30]:
    end_time = 'join_ts2_'+str(t)
    df[end_time] = df['Date'] + datetime.timedelta(t)

for entity in entities:
    data3 = df[['Recnum','Date', entity]]
    data4 = df[[ 'Recnum','join_ts1','join_ts2_0','join_ts2_1','join_ts2_3','join_ts2_7','join_ts2_14','join_ts2_30', entity]]
    data3.Recnum = data3.Recnum.astype(int)
    data4.Recnum = data4.Recnum.astype(int)
    tempdf = pd.merge(data3, data4, left_on = entity, right_on = entity)
    for t in [0,1,3,7,14,30]:
        name_end_time = 'join_ts2_' + str(t)
        tempdf2 = tempdf[(tempdf['Date'] >= tempdf['join_ts1']) & (tempdf['Date'] <= tempdf[name_end_time])& (tempdf['Recnum_x']>=tempdf['Recnum_y'])]
        temp_name = str(entity) + '_' + 'count' + str(t) + '_'
        tempGroupby = tempdf2[['Recnum_x','Date']].groupby('Recnum_x')
        tempGroupby = getattr(tempGroupby, 'count')().add_prefix(temp_name)
        tempGroupby.index = tempGroupby.index.rename('Recnum')
        frequency = pd.merge(frequency, tempGroupby,left_index = True, right_index = True, how = 'left')

frequency.info()

frequency.head()

frequency.to_csv('Frequency Variables.csv')
