==> Credit Card Transaction Fraud Analytics/2. Variable Creation - Relative velocity.ipynb <==
import pandas as pd
import numpy as np

df=pd.read_excel('no_missing.xlsx', converters={'Merchnum': lambda x: str(x), 'Merch_zip': lambda x : str(x)})

df.head()

# Cardnum
# Merchnum
# Merch_description_part

# Count velocity

import datetime

finaldf1=df[['Recnum']].set_index('Recnum')
finaldf2=df[['Recnum']].set_index('Recnum')

df['ts1'] = df['Date']

for t in [0,1,3,7,14,30]:
    end_time = 'ts2_'+str(t)
    df[end_time] = df['Date'] + datetime.timedelta(t)

entities=['Cardnum','Merchnum','Merch_description_part']
for entity in entities:
    data1=df[['Recnum','Date',entity]]
    data2=df[['ts1','ts2_0','ts2_1','ts2_3','ts2_7','ts2_14','ts2_30','Amount','Recnum',entity]]
    data1.Recnum=data1.Recnum.astype(int)
    data2.Recnum=data2.Recnum.astype(int)
    tempdf=pd.merge(data1,data2,left_on=entity,right_on=entity)
    for t in [0,1,3,7,14,30]:
        name_end_time='ts2_'+str(t)
        tempdf2=tempdf[(tempdf['Date']>=tempdf['ts1'])&(tempdf['Date']<=tempdf[name_end_time])&(tempdf['Recnum_x']>=tempdf['Recnum_y'])]
        temp_name=str(entity)+'_'+'count'+str(t)+'_'
        tempGroupby=tempdf2[['Recnum_x','Amount']].groupby('Recnum_x')
        tempGroupby=getattr(tempGroupby,'count')().add_prefix(temp_name)
        tempGroupby.index=tempGroupby.index.rename('Recnum')
        finaldf1=pd.merge(finaldf1,tempGroupby,left_index=True,right_index=True,how='left')

finaldf1

entities=['Cardnum','Merchnum','Merch_description_part']
for entity in entities:
    data1=df[['Recnum','Date',entity]]
    data2=df[['ts1','ts2_0','ts2_1','ts2_3','ts2_7','ts2_14','ts2_30','Amount','Recnum',entity]]
    data1.Recnum=data1.Recnum.astype(int)
    data2.Recnum=data2.Recnum.astype(int)
    tempdf=pd.merge(data1,data2,left_on=entity,right_on=entity)
    for t in [0,1,3,7,14,30]:
        name_end_time='ts2_'+str(t)
        tempdf2=tempdf[(tempdf['Date']>=tempdf['ts1'])&(tempdf['Date']<=tempdf[name_end_time])&(tempdf['Recnum_x']>=tempdf['Recnum_y'])]
        temp_name=str(entity)+'_'+'sum'+str(t)+'_'
        tempGroupby=tempdf2[['Recnum_x','Amount']].groupby('Recnum_x')
        tempGroupby=getattr(tempGroupby,'sum')().add_prefix(temp_name)
        tempGroupby.index=tempGroupby.index.rename('Recnum')
        finaldf2=pd.merge(finaldf2,tempGroupby,left_index=True,right_index=True,how='left')

finaldf2

finaldf3=pd.merge(finaldf1,finaldf2,left_index=True,right_index=True,how='left')

finaldf3

finaldf=df[['Recnum']].set_index('Recnum')

days_numerator = ['0','1']
days_denominator = ['7','14','30']

for b in entities:
    for c in days_numerator:
        for d in days_denominator:
            finaldf[b+'_'+c+'_'+'count'+d+'_'+'count_ratio']=\
            finaldf3[b+'_'+'count'+c+'_Amount']/finaldf3[b+'_'+'count'+d+'_Amount']/int(d)
            finaldf[b+'_'+c+'_'+'sum'+d+'_'+'sum_ratio']=\
            finaldf3[b+'_'+'sum'+c+'_Amount']/finaldf3[b+'_'+'sum'+d+'_Amount']/int(d)

finaldf

finaldf.to_excel('Relative velocity.xlsx')
