==> Credit Card Transaction Fraud Analytics/2. Variable Creation - Risk Table.ipynb <==
import pandas as pd
import numpy as np
import datetime

df=pd.read_excel('no_missing.xlsx', converters={'Merchnum': lambda x: str(x), 'Merch_zip': lambda x : str(x)})

df.head()

df['dayofweek']=df['Date'].dt.dayofweek
day_week=df[['Fraud','dayofweek']][df['Recnum']<=84299].groupby(by='dayofweek').mean().reset_index()
day_week.columns=['dayofweek','day_week_risk']

day_week

df=pd.merge(df,day_week,left_on='dayofweek',right_on='dayofweek',how='left')

grouprisk=df[['Recnum','day_week_risk']].set_index('Recnum')

grouprisk

grouprisk.to_excel('Risk Table.xlsx')
