==> Credit Card Transaction Fraud Analytics/3. Merge Variable.ipynb <==
import pandas as pd
import numpy as np

df1=pd.read_csv('amount_variables.csv',index_col=0)
df2=pd.read_csv('Days-since Variables.csv',index_col=0)
df3=pd.read_csv('Frequency Variables.csv',index_col=0)
df4=pd.read_excel('Relative velocity.xlsx',index_col=0)
df5=pd.read_excel('Risk Table.xlsx',index_col=0)
df6=pd.read_excel('no_missing.xlsx')

df1

df2

df3

df4

df5

df=pd.merge(df1,df2,left_index=True,right_index=True,how='left')
df=pd.merge(df,df3,left_index=True,right_index=True,how='left')
df=pd.merge(df,df4,left_index=True,right_index=True,how='left')
df=pd.merge(df,df5,left_index=True,right_index=True,how='left')

df.head()

rand=np.random.rand(96397)
df['rand']=rand

df6 = df6[['Fraud','Recnum']]

df6 = df6.set_index('Recnum')

df=pd.merge(df,df6,left_index=True,right_index=True,how='left')

df7=pd.read_csv('Benford.csv',index_col=0)

df7.head()

df=pd.merge(df,df7,left_index=True,right_index=True,how='left')

df.head()

df.to_csv('final_variable.csv')
