==> Credit Card Transaction Fraud Analytics/5. Final_variables.ipynb <==
import pandas as pd
import numpy as np

v = pd.read_csv('final_variable.csv')

selected_col = [
'Recnum',
'Merch_description_part_count0__max',
 'Merchnum_count0__max',
 'Merchnum_count1__max',
 'Merchnum_count1__sum',
 'Merchnum_count3__max',
 'cardnum-merch_description_part_count14__max',
 'cardnum-merch_description_part_count14__sum',
 'cardnum-merch_description_part_count30__max',
 'cardnum-merch_description_part_count3__sum',
 'cardnum-merch_state_count14__max',
 'cardnum-merch_state_count30__max',
 'cardnum-merch_state_merch_zip_count30__max',
 'cardnum-merch_state_merch_zip_count3__sum',
 'cardnum-merch_zip_count14__max',
 'cardnum-merchnum_count14__max',
 'cardnum-merchnum_count14__sum',
 'cardnum-merchnum_count1__max',
 'cardnum-merchnum_count30__max',
 'cardnum-merchnum_count3__max',
 'cardnum-merchnum_count3__sum',
'Fraud']

df = v[selected_col]

selected = ['Merch_description_part_count0__max',
 'Merchnum_count0__max',
 'Merchnum_count1__max',
 'Merchnum_count1__sum',
 'Merchnum_count3__max',
 'cardnum-merch_description_part_count14__max',
 'cardnum-merch_description_part_count14__sum',
 'cardnum-merch_description_part_count30__max',
 'cardnum-merch_description_part_count3__sum',
 'cardnum-merch_state_count14__max',
 'cardnum-merch_state_count30__max',
 'cardnum-merch_state_merch_zip_count30__max',
 'cardnum-merch_state_merch_zip_count3__sum',
 'cardnum-merch_zip_count14__max',
 'cardnum-merchnum_count14__max',
 'cardnum-merchnum_count14__sum',
 'cardnum-merchnum_count1__max',
 'cardnum-merchnum_count30__max',
 'cardnum-merchnum_count3__max',
 'cardnum-merchnum_count3__sum']

df.head()

independent = df[selected]

from sklearn.preprocessing import StandardScaler
scaler=StandardScaler()
data=scaler.fit_transform(independent.to_numpy())
data=pd.DataFrame(data,columns=independent.columns)

for col in list(data.columns):
    a = np.array(data[col].values.tolist())
    data[col] = np.where(a > 5, 5, a).tolist()

data['Recnum'] = df['Recnum']
data['Fraud'] = df['Fraud']

data.info()

data.to_csv('final_selected_variables.csv')

data.describe()
